feat: add detailed feedback in cross-check-pairs table
