+++ client/src/modules/CrossCheckPairsTable/pages/CrossCheckPairsTable/index.ts
@@ -1,0 +1,1 @@
+export { default as CrossCheckPairsTable } from './CrossCheckPairsTable';
