+++ nestjs/src/courses/cross-checks/dto/index.ts
@@ -1,2 +1,3 @@
 export * from './check-tasks-pairs.dto';
 export * from './cross-check-feedback.dto';
+export * from './cross-check-criteria-data.dto';
